+++ shelby.R
@@ -75,14 +75,12 @@ cases_by_age %>%
   geom_col(mapping = aes(x = date, y = total_cases),
            fill = "#a6192e") +
   facet_wrap(~ age_range)+
-  theme_gray(base_size = 16)+
+  theme_gray(base_size = 12)+
   theme(axis.text.x=element_text(size=rel(.8)))
 ggsave("figs/case data.png", height = 8, width = 12, units="in", dpi=600)
 
 # implications
   # having public data available
 
-<<<<<<< HEAD
 # make sure to insert a picture
-=======
->>>>>>> 48e49a08a786267161fe33362cbf956e73244972
+