+++ changes.md
@@ -5,3 +5,11 @@
 -   edited Abstract
 
 -   downloaded urbnmapr
+
+-   found a vax dataset with counties from MN
+
+-   adjusted the multiple graphs case data to display cases from beginning of Dec 2021 to end of Jan 2022
+
+-   also changed the color on the graph created above
+
+-   created the .md doc that will contain all the code and inserted the image of the graph